+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.4.7] - 2025-06-30
-- fix #24, add sci.h as class implementation for scientific and engineering notation.
-- add experimental folder
+- fix #24, add printHelpersMT.h, 
+  - class implementation for multThreading.
+  - scieng(), sci() and eng()
+- add MT examples
 - fix PRINTHELPERS_LIB_VERSION
+- obsolete PRINTHELPERS_VERSION
 - update readme.md
 - minor edits